Conflicts:
	src/java/nxt/NxtDbVersion.java
	test/java/nxt/BlockchainTest.java
	test/java/nxt/GeneratorTest.java
	test/java/nxt/http/LeaseTest.java
	test/java/nxt/http/MessageEncryptionTest.java
	test/java/nxt/http/PaymentAndMessagesSuite.java
	test/java/nxt/http/SendMessageTest.java
	test/java/nxt/http/SendMoneyTest.java

+++ src/java/nxt/peer/Peers.java
@@ -1,3 +1,19 @@
+/******************************************************************************
+ * Copyright © 2013-2015 The Nxt Core Developers.                             *
+ *                                                                            *
+ * See the AUTHORS.txt, DEVELOPER-AGREEMENT.txt and LICENSE.txt files at      *
+ * the top-level directory of this distribution for the individual copyright  *
+ * holder information and the developer policies on copyright and licensing.  *
+ *                                                                            *
+ * Unless otherwise agreed in a custom licensing agreement, no part of the    *
+ * Nxt software, including this file, may be copied, modified, propagated,    *
+ * or distributed except according to the terms contained in the LICENSE.txt  *
+ * file.                                                                      *
+ *                                                                            *
+ * Removal or modification of this copyright notice is prohibited.            *
+ *                                                                            *
+ ******************************************************************************/
+
 package nxt.peer;
 
 import nxt.Account;
@@ -32,8 +48,10 @@ import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.PriorityQueue;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -189,7 +207,8 @@ public final class Peers {
 
         maxNumberOfInboundConnections = Nxt.getIntProperty("nxt.maxNumberOfInboundConnections");
         maxNumberOfOutboundConnections = Nxt.getIntProperty("nxt.maxNumberOfOutboundConnections");
-        maxNumberOfConnectedPublicPeers = Nxt.getIntProperty("nxt.maxNumberOfConnectedPublicPeers");
+        maxNumberOfConnectedPublicPeers = Math.min(Nxt.getIntProperty("nxt.maxNumberOfConnectedPublicPeers"),
+                maxNumberOfOutboundConnections);
         maxNumberOfKnownPeers = Nxt.getIntProperty("nxt.maxNumberOfKnownPeers");
         minNumberOfKnownPeers = Nxt.getIntProperty("nxt.minNumberOfKnownPeers");
         connectTimeout = Nxt.getIntProperty("nxt.connectTimeout");
@@ -214,10 +233,12 @@ public final class Peers {
             ThreadPool.runBeforeStart(new Runnable() {
 
                 private void loadPeers(Collection<String> addresses) {
+                    int now = Nxt.getEpochTime();
                     for (final String address : addresses) {
                         Future<String> unresolvedAddress = peersService.submit(() -> {
                             PeerImpl peer = Peers.findOrCreatePeer(address, true);
                             if (peer != null) {
+                                peer.setLastUpdated(now);
                                 Peers.addPeer(peer);
                                 return null;
                             }
@@ -231,10 +252,23 @@ public final class Peers {
                 public void run() {
                     loadPeers(wellKnownPeers);
                     if (usePeersDb) {
-                        loadPeers(defaultPeers);
                         Logger.logDebugMessage("Loading known peers from the database...");
+                        loadPeers(defaultPeers);
                         if (savePeers) {
-                            loadPeers(PeerDb.loadPeers());
+                            List<PeerDb.Entry> dbPeers = PeerDb.loadPeers();
+                            for (PeerDb.Entry dbPeer : dbPeers) {
+                                Future<String> unresolvedAddress = peersService.submit(() -> {
+                                    PeerImpl peer = Peers.findOrCreatePeer(dbPeer.getAddress(), true);
+                                    if (peer != null) {
+                                        if (peer.getLastUpdated() == 0)
+                                            peer.setLastUpdated(dbPeer.getLastUpdated());
+                                        Peers.addPeer(peer);
+                                        return null;
+                                    }
+                                    return dbPeer.getAddress();
+                                });
+                                unresolvedPeers.add(unresolvedAddress);
+                            }
                         }
                     }
                 }
@@ -442,10 +476,7 @@ public final class Peers {
             getPeersRequest = JSON.prepareRequest(request);
         }
 
-        private volatile boolean addedNewPeer;
-        {
-            Peers.addListener(peer -> addedNewPeer = true, Event.NEW_PEER);
-        }
+        private volatile boolean updatedPeer;
 
         @Override
         public void run() {
@@ -466,9 +497,14 @@ public final class Peers {
                     JSONArray peers = (JSONArray)response.get("peers");
                     Set<String> addedAddresses = new HashSet<>();
                     if (peers != null) {
+                        int now = Nxt.getEpochTime();
                         for (Object announcedAddress : peers) {
                             PeerImpl newPeer = findOrCreatePeer((String) announcedAddress, true);
                             if (newPeer != null) {
+                                if (now - newPeer.getLastUpdated() > 24 * 3600) {
+                                    newPeer.setLastUpdated(now);
+                                    updatedPeer = true;
+                                }
                                 Peers.addPeer(newPeer);
                                 addedAddresses.add((String) announcedAddress);
                                 if (hasTooManyKnownPeers()) {
@@ -476,9 +512,9 @@ public final class Peers {
                                 }
                             }
                         }
-                        if (savePeers && addedNewPeer) {
+                        if (savePeers && updatedPeer) {
                             updateSavedPeers();
-                            addedNewPeer = false;
+                            updatedPeer = false;
                         }
                     }
 
@@ -508,20 +544,52 @@ public final class Peers {
         }
 
         private void updateSavedPeers() {
-            Set<String> oldPeers = new HashSet<>(PeerDb.loadPeers());
-            Set<String> currentPeers = Peers.peers.values().parallelStream().unordered()
-                    .filter(peer -> peer.getAnnouncedAddress() != null && !peer.isBlacklisted())
-                    .map(Peer::getAnnouncedAddress)
-                    .collect(Collectors.toSet());
-            Set<String> toDelete = new HashSet<>(oldPeers);
-            toDelete.removeAll(currentPeers);
+            int now = Nxt.getEpochTime();
+            //
+            // Load the current database entries and map the announced address to database entry
+            //
+            List<PeerDb.Entry> oldPeers = PeerDb.loadPeers();
+            Map<String, PeerDb.Entry> oldMap = new HashMap<>();
+            oldPeers.forEach(entry -> oldMap.put(entry.getAddress(), entry));
+            //
+            // Create the current peer map (note that there can be duplicate peer entries with
+            // the same announced address)
+            //
+            Map<String, PeerDb.Entry> currentPeers = new HashMap<>();
+            Peers.peers.values().forEach(peer -> {
+                if (peer.getAnnouncedAddress() != null && !peer.isBlacklisted() && now - peer.getLastUpdated() < 7*24*3600)
+                    currentPeers.put(peer.getAnnouncedAddress(), new PeerDb.Entry(peer.getAnnouncedAddress(), peer.getLastUpdated()));
+            });
+            //
+            // Build toDelete, toUpdate and toAdd lists
+            //
+            List<PeerDb.Entry> toDelete = new ArrayList<>(oldPeers.size());
+            oldPeers.forEach(entry -> {
+                if (currentPeers.get(entry.getAddress()) == null)
+                    toDelete.add(entry);
+            });
+            List<PeerDb.Entry> toUpdate = new ArrayList<>(currentPeers.size());
+            List<PeerDb.Entry> toAdd = new ArrayList<>(currentPeers.size());
+            currentPeers.values().forEach(entry -> {
+                PeerDb.Entry oldEntry = oldMap.get(entry.getAddress());
+                if (oldEntry == null)
+                    toAdd.add(entry);
+                else if (entry.getLastUpdated() - oldEntry.getLastUpdated() > 24*3600)
+                    toUpdate.add(entry);
+            });
+            //
+            // Nothing to do if all of the lists are empty
+            //
+            if (toDelete.isEmpty() && toUpdate.isEmpty() && toAdd.isEmpty())
+                return;
+            //
+            // Update the peer database
+            //
             try {
                 Db.db.beginTransaction();
                 PeerDb.deletePeers(toDelete);
-	            //Logger.logDebugMessage("Deleted " + toDelete.size() + " peers from the peers database");
-                currentPeers.removeAll(oldPeers);
-                PeerDb.addPeers(currentPeers);
-	            //Logger.logDebugMessage("Added " + currentPeers.size() + " peers to the peers database");
+                PeerDb.updatePeers(toUpdate);
+                PeerDb.addPeers(toAdd);
                 Db.db.commitTransaction();
             } catch (Exception e) {
                 Db.db.rollbackTransaction();
@@ -613,15 +681,12 @@ public final class Peers {
     }
 
     public static boolean hasTooManyInboundPeers() {
-        return peers.values().parallelStream().unordered()
-                .filter(Peer::isInbound)
-                .count() >= maxNumberOfInboundConnections;
+        return getPeers(Peer::isInbound, maxNumberOfInboundConnections).size() >= maxNumberOfInboundConnections;
     }
 
     public static boolean hasTooManyOutboundConnections() {
-        return peers.values().parallelStream().unordered()
-                .filter(peer -> !peer.isBlacklisted() && peer.getState() == Peer.State.CONNECTED && peer.getAnnouncedAddress() != null)
-                .count() >= maxNumberOfOutboundConnections;
+        return getPeers(peer -> !peer.isBlacklisted() && peer.getState() == Peer.State.CONNECTED && peer.getAnnouncedAddress() != null,
+                maxNumberOfOutboundConnections).size() >= maxNumberOfOutboundConnections;
     }
 
     public static PeerImpl findOrCreatePeer(String announcedAddress, boolean create) {
